Version 0.0.4.20:
- Unterstützung für C++-Userprogramme in user/other_userprogs
- Neues Userprogramm: Editor (C++). Kann bislang nur Dateien anzeigen (absoluten Pfad als erstes Kommandozeilenargument übergeben)
- -ffreestanding sollte nicht für Userprogramme verwendet werden (Ermöglicht C++-Userprogramme mit Clang)
- fgetc() sollte EOF zurückgeben, falls das Dateiende erreicht wurde
- PRINT+h (Screenshot auf Harddisk) dokumentiert
- Hacks in der Umgebung von saveScreenshot etwas verringert

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1405 31957799-4632-4615-9793-72db6d48bcac

+++ user/other_userprogs/calc.c
@@ -156,7 +156,7 @@ static int32_t CalcTerm(char* term, size_t pos) {
     return (atoi(term+pos));
 }
 
-int main(size_t argc, char* argv[]) {
+int main(int argc, char* argv[]) {
     if(argc > 1)
     {
         printf("Result: %u", CalcTerm(argv[1], 0));
@@ -184,7 +184,7 @@ int main(size_t argc, char* argv[]) {
 }
 
 /*
-* Copyright (c) 2011 The PrettyOS Project. All rights reserved.
+* Copyright (c) 2011-2013 The PrettyOS Project. All rights reserved.
 *
 * http://www.c-plusplus.de/forum/viewforum-var-f-is-62.html
 *
